use the correct attribute mapping for the eduperson_principal_name field, and fix username generation
The oidc claim from which the username is extracted (eppn) is no longer mapped to a User field (as voperson_external_id is now put into the eduperson_principal_name field).  Therefore, we need to pass the oidc claim to generate_unique_username in order to extract the username.

Sooner or later, we will have to fix the mess of mismatching User field names and oidc claims.