Update collaboration_invitation.html
After user feedback where a % of users seem confused about what link to click, tried to clarify some more what link to click. Complicating in this case is, besides the fact we're offering several links in the email (all with good intend), we also offer the inviter to add a text, and they also sometimes add links to their message.